backend/requirements.txt
fastapi
uvicorn
elasticsearch
transformers
pymongo
tensorflow
tf-keras
numpy
torch
googlesearch-python
requests
sentencepiece
google-generativeai